wandermind-backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
anthropic==0.25.0
Pillow==10.3.0
requests==2.31.0
celery==5.3.6
redis==5.0.4
psycopg2-binary==2.9.9
django-ratelimit==4.1.0
drf-spectacular==0.27.2
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
